rename dist_reduce to teardown; compile gradient processing; move builder to a standalone file
